% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_scores.R
\name{crps.vis.wmo}
\alias{crps.vis.wmo}
\title{CRPS of WMO-reported visibility forecasts based on the estimated probabilities
and verifying observations}
\usage{
crps.vis.wmo(prob, obs)
}
\arguments{
\item{prob}{Estimated probabilities. Dimensions: \code{[84, 21]}, where \code{84} is the number of observation categories and 21 is the number of lead times.}

\item{obs}{A \code{[1:21]}-sized array, containing the verifying observations.}
}
\value{
CRPS for a specific location.
}
\description{
Calculates the CRPS
}
