% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_scores.R
\name{logs.vis.wmo}
\alias{logs.vis.wmo}
\title{Logarithmic Score of WMO-reported visibility forecasts based on the estimated probabilities
and verifying observations}
\usage{
logs.vis.wmo(prob, obs)
}
\arguments{
\item{prob}{Estimated probabilities. Dimensions: \code{[84, 21]}, where \code{84} is the number of observation categories and 21 is the number of lead times.}

\item{obs}{A \code{[1:21]}-sized array, containing the verifying observations.}
}
\value{
Logarithmic score for a specific location
}
\description{
Logarithmic Score of WMO-reported visibility forecasts based on the estimated probabilities
and verifying observations
}
