% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_scores.R
\name{pit.vis.wmo}
\alias{pit.vis.wmo}
\title{PIT values based on the estimated probabilities
and verifying observations}
\usage{
pit.vis.wmo(prob, obs)
}
\arguments{
\item{obs}{Verifying observations. Dimensions:}

\item{probs}{Estimated probabilities. Dimensions:}
}
\value{
PIT values
}
\description{
PIT values based on the estimated probabilities
and verifying observations
}
