% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{polrProbsVis}
\alias{polrProbsVis}
\title{Fitting POLR-R model to the EUPPBench visibility ensemble}
\usage{
polrProbsVis(
  visData,
  varNames,
  nForcDates,
  nStats,
  nObsNames,
  nLeadTimes,
  forcDates,
  stationIDs,
  obsNames,
  leadTimes,
  start.day,
  end.day,
  train = 30,
  startZeta = seq(-1.7, 6.5, by = 0.1)
)
}
\arguments{
\item{visData}{Visibility data. Dimensions:}

\item{varNames}{Feature names}

\item{nForcDates}{Length of verification period}

\item{nStats}{Number of stations}

\item{nObsNames}{Number of observation names}

\item{nLeadTimes}{Number of lead times}

\item{forcDates}{A character vector of dates in the verification period}

\item{stationIDs}{A character vector of station ids.}

\item{obsNames}{A character vector of observation names}

\item{leadTimes}{A character vector of lead times}

\item{start.day}{Index of the first day of the verification period}

\item{end.day}{Index of the last day of the verification period}

\item{train}{Length of training period (default value is 30 days)}

\item{startZeta}{}
}
\value{
Estimated probabilities
}
\description{
Fitting POLR-R model to the EUPPBench datasets' visibility forecasts.
Architecture: 2 hidden layers (25x25 neurons), Activation function: Logistic
Input features:
\itemize{
\item control forecast (normalized)
\item ensemble mean (normalized)
\item ensemble variance (normalized)
\item proportions of forecasts predicting visibility between 0 and 1000 meters
\item proportions of forecasts predicting visibility between 1001 and 2000 meters
\item proportions of forecasts predicting visibility above 30000 meters
}
}
