% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_scores.R
\name{quant.vis.wmo}
\alias{quant.vis.wmo}
\title{Quantiles of the discrete probability distribution of WMO-reported visibility forecasts}
\usage{
quant.vis.wmo(prob, q)
}
\arguments{
\item{prob}{Estimated probabilities. Dimensions: \code{[84, 21]}, where \code{84} is the number of observation categories and 21 is the number of lead times.}

\item{q}{level of quantile}
}
\value{
Estimated quantiles of the discrete distribution for the 21 different lead times
}
\description{
Quantiles of the discrete probability distribution of WMO-reported visibility forecasts
}
