% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode.wmo}
\alias{recode.wmo}
\title{Recoding raw visibility ensemble forecasts according to the reported observation
categories of the WMO}
\usage{
recode.wmo(ens)
}
\arguments{
\item{ens}{Raw ensemble forecasts}
}
\value{
Recoded ensemble forecast
}
\description{
Recoding raw visibility ensemble forecasts according to the reported observation
categories of the WMO
}
