% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_scores.R
\name{var.vis.wmo}
\alias{var.vis.wmo}
\title{Estimated variance of the discrete probability distribution of WMO-reported visibility forecasts}
\usage{
var.vis.wmo(prob)
}
\arguments{
\item{prob}{Estimated probabilities. Dimensions: \code{[84, 21]}, where \code{84} is the number of observation categories and 21 is the number of lead times.}
}
\value{
Estimated variances for the 21 different lead times
}
\description{
Estimated variance of the discrete probability distribution of WMO-reported visibility forecasts
}
